Moved code from convert_temp to convert.
